requirements-dashboard.txt
# === Memecoin Gem Detector - Dashboard Dependencies ===
# Reduced set for Docker/Render deployment (dashboard only)

# --- Data & ML ---
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
joblib>=1.3

# --- Visualization ---
plotly>=5.15

# --- Dashboard ---
streamlit>=1.28

# --- HTTP & APIs ---
requests>=2.31

# --- Config ---
python-dotenv>=1.0

# --- Storage ---
pyarrow>=14.0
supabase>=2.0
